check key registration on processing queue empty
