+++ src/Nethermind/Nethermind.Shutter/ShutterBlockHandler.cs
@@ -4,10 +4,7 @@
 using System;
 using System.Collections.Generic;
 using Nethermind.Core;
-using Nethermind.Evm.TransactionProcessing;
-using Nethermind.Shutter.Contracts;
 using Nethermind.Logging;
-using Nethermind.Abi;
 using Nethermind.Blockchain.Receipts;
 using System.Threading.Tasks;
 using System.Threading;
@@ -15,7 +12,6 @@ using Nethermind.Core.Caching;
 using Nethermind.Core.Crypto;
 using Nethermind.Blockchain;
 using Nethermind.Core.Collections;
-using Nethermind.Shutter.Config;
 
 namespace Nethermind.Shutter;
 
@@ -26,30 +22,18 @@ public class ShutterBlockHandler : IShutterBlockHandler
     private readonly IShutterEon _eon;
     private readonly IReceiptFinder _receiptFinder;
     private readonly ShutterTxLoader _txLoader;
-    private readonly ShutterValidatorsInfo _validatorsInfo;
-    private readonly ILogManager _logManager;
-    private readonly IAbiEncoder _abiEncoder;
     private readonly IBlockTree _blockTree;
     private readonly ReadOnlyBlockTree _readOnlyBlockTree;
-    private readonly ulong _chainId;
-    private readonly IShutterConfig _cfg;
     private readonly TimeSpan _slotLength;
     private readonly TimeSpan _blockWaitCutoff;
-    private readonly IShareableTxProcessorSource _txProcessorSource;
-    // private bool _haveCheckedRegistered = false;
     private ulong _blockWaitTaskId = 0;
     private readonly Dictionary<ulong, Dictionary<ulong, BlockWaitTask>> _blockWaitTasks = [];
     private readonly LruCache<ulong, Hash256?> _slotToBlockHash = new(5, "Slot to block hash mapping");
     private readonly Lock _syncObject = new();
 
     public ShutterBlockHandler(
-        ulong chainId,
-        IShutterConfig cfg,
-        IShareableTxProcessorSource txProcessorSource,
         IBlockTree blockTree,
-        IAbiEncoder abiEncoder,
         IReceiptFinder receiptFinder,
-        ShutterValidatorsInfo validatorsInfo,
         IShutterEon eon,
         ShutterTxLoader txLoader,
         SlotTime time,
@@ -57,19 +41,13 @@ public class ShutterBlockHandler : IShutterBlockHandler
         TimeSpan slotLength,
         TimeSpan blockWaitCutoff)
     {
-        _chainId = chainId;
-        _cfg = cfg;
         _logger = logManager.GetClassLogger();
         _time = time;
-        _validatorsInfo = validatorsInfo;
         _eon = eon;
         _receiptFinder = receiptFinder;
         _txLoader = txLoader;
         _blockTree = blockTree;
         _readOnlyBlockTree = blockTree.AsReadOnly();
-        _abiEncoder = abiEncoder;
-        _logManager = logManager;
-        _txProcessorSource = txProcessorSource;
         _slotLength = slotLength;
         _blockWaitCutoff = blockWaitCutoff;
 
@@ -159,11 +137,6 @@ public class ShutterBlockHandler : IShutterBlockHandler
         {
             if (_logger.IsDebug) _logger.Debug($"Shutter block handler {head.Number}");
 
-            // if (!_haveCheckedRegistered)
-            // {
-            //     CheckAllValidatorsRegistered(head.Header, _validatorsInfo);
-            //     _haveCheckedRegistered = true;
-            // }
             _eon.Update(head.Header);
             _txLoader.LoadFromReceipts(head, _receiptFinder.Get(head), _eon.GetCurrentEonInfo()!.Value.Eon);
 
@@ -185,28 +158,6 @@ public class ShutterBlockHandler : IShutterBlockHandler
         else if (_logger.IsDebug)
         {
             _logger.Warn($"Shutter block handler not running, outdated block {head.Number}");
-        }
-    }
-
-
-    private void CheckAllValidatorsRegistered(in BlockHeader parent, in ShutterValidatorsInfo validatorsInfo)
-    {
-        if (validatorsInfo.IsEmpty)
-        {
-            return;
-        }
-
-        using IReadOnlyTxProcessingScope scope = _txProcessorSource.Build(parent);
-        ITransactionProcessor processor = scope.TransactionProcessor;
-
-        ValidatorRegistryContract validatorRegistryContract = new(processor, _abiEncoder, new(_cfg.ValidatorRegistryContractAddress!), _logManager, _chainId, _cfg.ValidatorRegistryMessageVersion!);
-        if (validatorRegistryContract.IsRegistered(parent, validatorsInfo, out HashSet<ulong> unregistered))
-        {
-            if (_logger.IsInfo) _logger.Info($"All Shutter validator keys are registered.");
-        }
-        else if (_logger.IsError)
-        {
-            _logger.Error($"Validators not registered to Shutter with the following indices: [{string.Join(", ", unregistered)}]");
         }
     }
